# docs/openapi.yaml
openapi: 3.0.0
info:
  title: Bank REST API
  version: 1.0.0
servers:
  - url: http://localhost:8080
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      properties:
        timestamp: {type: string, format: date-time}
        status: {type: integer}
        error: {type: string}
        message: {type: string}
    PageCardDTO:
      type: object
      properties:
        content:
          type: array
          items: {$ref: '#/components/schemas/CardDTO'}
        totalElements: {type: integer}
        totalPages: {type: integer}
        size: {type: integer}
        number: {type: integer}
    CardDTO:
      type: object
      properties:
        id: {type: integer}
        cardNumber: {type: string, description: '**** **** **** 1234'}
        cardHolderName: {type: string}
        expirationDate: {type: string, format: date}
        status: {type: string, enum: [ACTIVE, BLOCKED, EXPIRED]}
        balance: {type: number, format: double}
        ownerId: {type: integer}
        ownerUsername: {type: string}
    LoginRequest:
      type: object
      required: [username, password]
      properties: {username: {type: string}, password: {type: string}}
    RefreshTokenRequest:
      type: object
      required: [refreshToken]
      properties: {refreshToken: {type: string}}
    JwtResponse:
      type: object
      properties:
        token: {type: string}
        refreshToken: {type: string}
        type: {type: string}
        username: {type: string}
        roles: {type: array, items: {type: string}}
    CardCreateRequest:
      type: object
      required: [cardNumber, cardHolderName, expirationDate]
      properties:
        cardNumber: {type: string}
        cardHolderName: {type: string}
        expirationDate: {type: string, format: date}
        ownerId: {type: integer}
    CardUpdateRequest:
      type: object
      properties:
        cardHolderName: {type: string}
        expirationDate: {type: string, format: date}
        status: {type: string, enum: [ACTIVE, BLOCKED, EXPIRED]}
        balance: {type: number, format: double}
        ownerId: {type: integer}
security:
  - bearerAuth: []
paths:
  /api/auth/login:
    post:
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/LoginRequest'}
      responses:
        '200': {description: OK, content: {application/json: {schema: {$ref: '#/components/schemas/JwtResponse'}}}}
        '401': {description: Unauthorized, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
      security: []
  /api/auth/refresh:
    post:
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/RefreshTokenRequest'}
      responses:
        '200': {description: OK, content: {application/json: {schema: {$ref: '#/components/schemas/JwtResponse'}}}}
        '401': {description: Unauthorized, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
      security: []
  /api/cards:
    get:
      summary: Get user's cards
      parameters:
        - in: query
          name: page
          schema: {type: integer, default: 0}
        - in: query
          name: size
          schema: {type: integer, default: 10}
        - in: query
          name: sortBy
          schema: {type: string}
        - in: query
          name: search
          schema: {type: string}
      responses:
        '200': {description: OK, content: {application/json: {schema: {$ref: '#/components/schemas/PageCardDTO'}}}}
        '401': {description: Unauthorized}
        '403': {description: Forbidden}
    post:
      summary: Create card
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/CardCreateRequest'}}}
      responses:
        '201': {description: Created, content: {application/json: {schema: {$ref: '#/components/schemas/CardDTO'}}}}
        '400': {description: Bad Request, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
        '401': {description: Unauthorized}
        '403': {description: Forbidden}
  /api/admin/cards:
    get:
      summary: Get all cards (admin)
      parameters:
        - in: query
          name: page
          schema: {type: integer, default: 0}
        - in: query
          name: size
          schema: {type: integer, default: 10}
        - in: query
          name: sortBy
          schema: {type: string}
        - in: query
          name: search
          schema: {type: string}
      responses:
        '200': {description: OK, content: {application/json: {schema: {$ref: '#/components/schemas/PageCardDTO'}}}}
        '401': {description: Unauthorized}
        '403': {description: Forbidden}
  /api/cards/{id}:
    get:
      summary: Get card by id (owner)
      parameters: [{in: path, name: id, required: true, schema: {type: integer}}]
      responses:
        '200': {description: OK, content: {application/json: {schema: {$ref: '#/components/schemas/CardDTO'}}}}
        '401': {description: Unauthorized}
        '403': {description: Forbidden}
    put:
      summary: Update card (admin)
      parameters: [{in: path, name: id, required: true, schema: {type: integer}}]
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/CardUpdateRequest'}}}
      responses:
        '200': {description: OK, content: {application/json: {schema: {$ref: '#/components/schemas/CardDTO'}}}}
        '400': {description: Bad Request}
        '401': {description: Unauthorized}
        '403': {description: Forbidden}
    delete:
      summary: Delete card (admin)
      parameters: [{in: path, name: id, required: true, schema: {type: integer}}]
      responses:
        '204': {description: No Content}
        '401': {description: Unauthorized}
        '403': {description: Forbidden}
  /api/cards/{id}/block:
    put:
      summary: Block card (owner/admin)
      parameters: [{in: path, name: id, required: true, schema: {type: integer}}]
      responses:
        '200': {description: OK, content: {application/json: {schema: {$ref: '#/components/schemas/CardDTO'}}}}
        '401': {description: Unauthorized}
        '403': {description: Forbidden}
  /api/cards/{id}/activate:
    put:
      summary: Activate card (admin)
      parameters: [{in: path, name: id, required: true, schema: {type: integer}}]
      responses:
        '200': {description: OK, content: {application/json: {schema: {$ref: '#/components/schemas/CardDTO'}}}}
        '400': {description: Bad Request}
        '401': {description: Unauthorized}
        '403': {description: Forbidden}
  /api/transfers:
    post:
      summary: Transfer between own cards
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/TransferRequest'}}}
      responses:
        '200': {description: OK}
        '400': {description: Bad Request}
        '401': {description: Unauthorized}
        '403': {description: Forbidden}